feat: Use ninja for the repo maintenance items
For basic maintenance tasks that have dependencies, I have had great
success using basic ninja rules to stitch them together; leads to a
simpler setup and better development experience.

Note that the nixpkgs update script has to be run manually, since it’s
the base dependency that we depend on for everything (and we need to
be in the nix shell for changes to be applied).