fix(watcher): rephrase macos event drop test fix
`macos_eat_late_notifications` is now less straightforward to
implement, because we separated `Watch` from `Filter` and don’t expose
the filter channel anymore.

So instead let’s pass an optional duration, during which the first
event is dropped.

This somewhat complicates the event loop, but it can’t be helped, I
think??