fix(watcher): only watch /bar to work around macos notify race
MacOS has some delay between any change and when the watcher actually
notices it. So it would always show the creation of foo when we want
to make sure `bar` does not get watched, leading to the test failing.

Instead, filter out every file that is not `/bar` here, which should
be fine.